[Refactor] Update server structure (#11)
* refactor: make some code composition changes

- renaming controllers to services (will make them less reliant on the res,req props)
- convert callback db functions to async

* refactor: decouple api handling from business logic


This PR refactors the server structure, separating the business logic from the api controllers. This will make the business logic independently testable (if necessary) and allow the logic called/used independently of the endpoints. This PR also replaced the existing database functions that used callbacks with a implementation that can be awaited.

---------

Co-authored-by: JChris246 <JChris246@users.noreply.github.com>